requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
lightgbm==4.6.0
shap==0.47.2
matplotlib==3.10.3
seaborn==0.13.2
streamlit==1.45.1
joblib==1.4.2
pillow==11.2.1

